+++ src/AtomUI.Desktop.Controls/Steps/StepsItemPseudoClass.cs
@@ -1,0 +1,8 @@
+namespace AtomUI.Desktop.Controls;
+
+internal static class StepsItemPseudoClass
+{
+    internal const string ItemHeaderForegroundOverride = ":item-header-foreground-override";
+    internal const string ItemSubHeaderForegroundOverride = ":item-sub-header-foreground-override";
+    internal const string ItemRailBackgroundOverride = ":item-rail-background-override";
+}